ESM Source Folder
The source folder is now build using TS as well! It's still output as CJS for the time being though. I have to fix the JSDoc type placement now that it no longer sets things to the module exports object, which make it more complex to type things out (it can't be declared in a comment above).