ESM + Tooling Upgrades
I'm not sure how to use or configure ESLint yet, I have yet to try it out myself. I'm going to upgrade their versions to the latest releases though.

This actually made things work very easily as ESM out of the box! Extra awesome! I thought I was going to have to configure a bunch more to get it working. Turns out Mocha supports it, it's Jest that doesn't. Or something along those lines when I tried it last. I think that was for JSMediaTags I was getting stuck with those kinds of issues.

I didn't end up using these links to set things up, but it helped me get to the right place. The plain dependencies fixed the issue itself, it seems.

https://github.com/mochajs/mocha/issues/4853
https://daveschinkel.medium.com/setting-up-mocha-js-to-run-typescript-42f3099450eb
https://stackblitz.com/edit/node-pawwmm
https://mochajs.org/#serial-mode
https://stackoverflow.com/questions/54805319/mocha-error-no-test-files-found-test-npm-err-test-failed