Simple Formatting
Just went through all of the source files and formatted them with VSCode, to make things more "standardized" if you will. This is probably my learning of Rust taking hold of me now. I have now joined the dark side, rustfmt is awesome! (I used to be a heavy opposer to opinionated formatting; it was because I didn't know any better 😉)